Added a new WS event for when the dashboard is updated. Moved some interfaces into their own files for better tree shaking. Preparing to refresh dashboard on WS events.
