+++ UI/Web/src/app/_models/dashboard/stream-type.enum.ts
@@ -1,0 +1,7 @@
+export enum StreamType {
+  OnDeck = 1,
+  RecentlyUpdated = 2,
+  NewlyAdded = 3,
+  SmartFilter = 4,
+  MoreInGenre = 5
+}
