+++ UI/Web/src/app/_models/events/dashboard-update-event.ts
@@ -1,0 +1,3 @@
+export interface DashboardUpdateEvent {
+
+}
