+++ UI/Web/src/app/sidenav/_components/customize-dashboard-modal/customize-dashboard-modal.component.html
@@ -12,7 +12,7 @@
     </app-draggable-ordered-list>
 
     <h5>Smart Filters</h5>
-    <ul class="filter-list">
+    <ul class="list-group filter-list">
       <li class="filter list-group-item" *ngFor="let filter of smartFilters">
         {{filter.name}}
         <button class="btn btn-icon" (click)="addFilterToStream(filter)">
@@ -27,7 +27,6 @@
   </div>
 
   <div class="modal-footer">
-    <button type="button" class="btn btn-secondary" (click)="close()">{{t('close')}}</button>
-    <button type="button" class="btn btn-primary" (click)="save()">{{t('save')}}</button>
+    <button type="button" class="btn btn-primary" (click)="close()">{{t('close')}}</button>
   </div>
 </ng-container>
